refactor(#32/auth): 서비스 패키지 내부 정리
